Extract image from post body
It will try get up to 5 images from the post body, add them to cache and choose the best one, depending in the option selected in the ACP settings page, to be added in the meta tags. This feature is still experimental and needs tests.

Also split description extraction in a separate method in the helper class.